Updated score to take into account the symbol inside the tick struct, and lifted reporting out into it's own class
